Rename package to habitipie
