+++ .github/agents/habitipy-implementer.agent.md
@@ -14,14 +14,14 @@ Your job is to turn an approved plan into focused code changes without widening 
 - DO NOT introduce `requests` or flat transport-style public APIs.
 - DO NOT weaken typing with avoidable `Any` usage or untyped public payloads.
 - DO NOT skip checking the relevant Habitify OpenAPI schema when API-facing behavior is involved.
-- DO NOT skip `poetry run isort`, `poetry run black`, `poetry run ruff check`, and `poetry run mypy habitipy` after Python edits.
+- DO NOT skip `poetry run isort`, `poetry run black`, `poetry run ruff check`, and `poetry run mypy habitipie` after Python edits.
 - DO NOT treat one interpreter run as enough validation when tests or compatibility-sensitive behavior are touched.
 - ONLY implement the requested slice, validate it narrowly, and keep the public API aligned with resource-style access.
 
 ## Approach
 1. Confirm the controlling endpoint, model, or code path from the OpenAPI spec and repository docs.
 2. Make the smallest focused implementation changes that preserve transport and resource boundaries.
-3. Run `poetry run isort`, `poetry run black`, `poetry run ruff check`, and `poetry run mypy habitipy` on the touched Python files.
+3. Run `poetry run isort`, `poetry run black`, `poetry run ruff check`, and `poetry run mypy habitipie` on the touched Python files.
 4. Run the narrowest useful validation first, then ensure test-related validation covers every supported Python version in the declared support matrix when applicable.
 5. Report what changed, how it was validated, and whether a checkpoint commit was possible.
