+++ .github/workflows/release.yml
@@ -1,18 +1,16 @@
-name: CI
+name: Release
 
 on:
-  pull_request:
   push:
-    branches:
-      - main
-  workflow_dispatch:
-
-concurrency:
-  group: ci-${{ github.workflow }}-${{ github.ref }}
-  cancel-in-progress: true
+    tags:
+      - "v*.*.*"
 
 permissions:
   contents: read
+
+concurrency:
+  group: release-${{ github.ref }}
+  cancel-in-progress: false
 
 jobs:
   quality:
@@ -69,15 +67,15 @@ jobs:
       - name: Install dependencies
         run: poetry install --extras dev
 
-      - name: Run tests in tox
+      - name: Run tests
         run: poetry run tox run -e ${{ matrix.tox-env }}
 
-  coverage:
-    name: Coverage
+  build:
+    name: Build distributions
+    needs:
+      - quality
+      - tests
     runs-on: ubuntu-latest
-    permissions:
-      contents: read
-      id-token: write
 
     steps:
       - name: Check out repository
@@ -88,18 +86,50 @@ jobs:
         with:
           python-version: "3.12"
 
-      - name: Install Poetry
-        run: python -m pip install poetry
+      - name: Check tag matches package version
+        run: |
+          VERSION=$(python - <<'PY'
+          import tomllib
+          with open("pyproject.toml", "rb") as f:
+              print(tomllib.load(f)["project"]["version"])
+          PY
+          )
+          test "${GITHUB_REF_NAME}" = "v${VERSION}"
 
-      - name: Install dependencies
-        run: poetry install --extras dev
+      - name: Install build tooling
+        run: python -m pip install --upgrade build twine
 
-      - name: Generate coverage report
-        run: poetry run pytest --cov=habitipie --cov-report=xml
+      - name: Build distributions
+        run: python -m build
 
-      - name: Upload coverage reports to Codecov
-        uses: codecov/codecov-action@v5
+      - name: Check distributions
+        run: python -m twine check dist/*
+
+      - name: Store distributions
+        uses: actions/upload-artifact@v4
         with:
-          token: ${{ secrets.CODECOV_TOKEN }}
-          files: ./coverage.xml
-          use_oidc: true
+          name: python-package-distributions
+          path: dist/
+
+  publish:
+    name: Publish to PyPI
+    needs:
+      - build
+    runs-on: ubuntu-latest
+
+    environment:
+      name: pypi
+      url: https://pypi.org/p/habitipie
+
+    permissions:
+      id-token: write
+
+    steps:
+      - name: Download distributions
+        uses: actions/download-artifact@v4
+        with:
+          name: python-package-distributions
+          path: dist/
+
+      - name: Publish distributions to PyPI
+        uses: pypa/gh-action-pypi-publish@release/v1
