+++ pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "habitipy"
+name = "habitipie"
 version = "0.1.0"
 description = "A Python client for Habitify: manage habits, logs, completions, skips, and progress from code."
 authors = [
@@ -44,7 +44,7 @@ target-version = ["py310"]
 profile = "black"
 line_length = 100
 py_version = 310
-src_paths = ["habitipy", "tests"]
+src_paths = ["habitipie", "tests"]
 
 
 [tool.ruff]
@@ -63,7 +63,7 @@ indent-style = "space"
 
 [tool.mypy]
 python_version = "3.10"
-files = ["habitipy"]
+files = ["habitipie"]
 plugins = ["pydantic.mypy"]
 check_untyped_defs = true
 disallow_any_generics = true
@@ -77,7 +77,7 @@ warn_unused_ignores = true
 
 
 [tool.coverage.run]
-source = ["habitipy"]
+source = ["habitipie"]
 
 
 [tool.coverage.report]
