+++ tox.ini
@@ -10,7 +10,7 @@ commands = pytest
 basepython = python3.12
 extras = dev
 commands =
-	isort --check-only habitipy tests
-	black --check habitipy tests
-	ruff check habitipy tests
-	mypy habitipy
+	isort --check-only habitipie tests
+	black --check habitipie tests
+	ruff check habitipie tests
+	mypy habitipie
